requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
supabase==2.0.1
requests==2.31.0
jwt==1.4.0